bot/requirements.txt
python-telegram-bot>=20.7
llama-cpp-python>=0.2.0
gql[aiohttp]>=3.5.0
graphql-core>=3.2.0
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
aiohttp>=3.9.0
jinja2>=3.1.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
